refactor(string): 기존 string sum 내부상태를 나타내던 클래스의 이름을 StringCalculator -> StringSumCalculation으로 변경
- StringCalculator 신설, static 메서드 사용가능
